+++ src/io/layouts/user/user_stats.hh
@@ -33,25 +33,25 @@ namespace shiro::io::layouts {
 
         std::string beatmap_checksum = "";
 
-        uint32_t current_mods = 0;
+        int64_t current_mods = 0;
         uint8_t play_mode = 0;
 
         int32_t beatmap_id = 0;
 
-        uint64_t ranked_score = 0;
+        int64_t ranked_score = 0;
 
-        uint64_t total_hits = 0;
+        int32_t total_hits = 0;
         int32_t max_combo = 0;
         float accuracy = 0.0f; // Between 0 and 1 (divide by 100 I guess)
         int32_t play_count = 0;
         int32_t play_time = 0;
-        uint64_t total_score = 0;
+        int64_t total_score = 0;
 
-        uint32_t count_A = 0;
-        uint32_t count_S = 0;
-        uint32_t count_X = 0;  // SS
-        uint32_t count_SH = 0; // S  with Hidden
-        uint32_t count_XH = 0; // SS with Hidden
+        int32_t count_A = 0;
+        int32_t count_S = 0;
+        int32_t count_X = 0;  // SS
+        int32_t count_SH = 0; // S  with Hidden
+        int32_t count_XH = 0; // SS with Hidden
 
         int32_t rank = 0;
         int16_t pp = 0;
